Implement peers support through config instead of relations
The old-style charm uses relation with another charm,
thruk-external-agent, to model the remote thruks. These apps are stubs,
whose only reason of existence is to hold the configuration needed for
the local thruk to interface with the remote one.

That approach is unfeasible in the k8s way, since it'd imply spinning up
an entire container to do nothing but sit there holding a few key/value
pairs in its juju agent.

Hence this approach using a dictionary in a config property was adopted.